# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a Gin-Gonic API with gRPC integration.",
        "title": "Gin-Gonic API with gRPC",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/comment": {
            "post": {
                "description": "新增留言",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "留言"
                ],
                "summary": "新增留言",
                "parameters": [
                    {
                        "description": "Create comment",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gin_service.Comment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin_service.httpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin_service.httpResponse"
                        }
                    }
                }
            }
        },
        "/comment/{id}": {
            "get": {
                "description": "獲取留言",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "留言"
                ],
                "summary": "獲取留言",
                "parameters": [
                    {
                        "type": "string",
                        "description": "comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin_service.httpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin_service.httpResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "刪除留言",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "留言"
                ],
                "summary": "刪除留言",
                "parameters": [
                    {
                        "type": "string",
                        "description": "comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin_service.httpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin_service.httpResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gin_service.Comment": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "1"
                },
                "message": {
                    "type": "string",
                    "example": "test msg"
                }
            }
        },
        "gin_service.httpResponse": {
            "type": "object",
            "properties": {
                "response": {
                    "type": "string"
                }
            }
        }
    }
}